refactor: make protocol limits and security profiles explicit
Replace mutable test defaults with explicit Limits carried through FROST and CGGMP21 plan options, validation, payload decoding, and artifact APIs. Keep local resource policy out of shared plan digests while preserving fail-closed production defaults.

Make internal/zk/paillier.SecurityParams the canonical uint32 definition and expose it from cggmp21/secp256k1 as an alias. Add canonical wire encoding, bind profiles into lifecycle digests, and persist them as nested records in key shares, presigns, and reshare plans. Remove the overlapping keygen PaillierBits setting while retaining refresh and reshare key-size selection.

Harden parameter validation against undersized/oversized moduli and proof-range overflow, return errors for reduced Paillier generation outside tests, and restore nil-safe reshare plan marshaling. Update ZK and MtA boundaries to use uint32 parameters directly.

Move reduced profiles and relaxed limits into package-local test helpers, expand profile persistence and encoding coverage, regenerate CGGMP21 wire/protocol vectors, add the SecurityParams golden, and update public examples and documentation.

+++ internal/zk/paillier/affg.go
@@ -610,10 +610,10 @@ func buildAffGTranscript(params SecurityParams, state []byte, stmt AffGStatement
 	t.AppendBytes("curve", []byte("secp256k1"))
 	t.AppendBytes("proof", []byte("aff-g"))
 	t.AppendUint16("version", 1)
-	t.AppendUint32("ell", uint32(params.Ell))
-	t.AppendUint32("ell_prime", uint32(params.EllPrime))
-	t.AppendUint32("epsilon", uint32(params.Epsilon))
-	t.AppendUint32("challenge_bits", uint32(params.ChallengeBits))
+	t.AppendUint32("ell", params.Ell)
+	t.AppendUint32("ell_prime", params.EllPrime)
+	t.AppendUint32("epsilon", params.Epsilon)
+	t.AppendUint32("challenge_bits", params.ChallengeBits)
 	t.AppendBytes("state", state)
 
 	// Statement.
